fear(API): Add MAM file to save configures
We could simply load the MAM file to get MAM
object which would save information about MSS
tree and channel ordinal.

So there are two way to set channel_ord in the current design:
1. Use MAM file which means no channel_ord field in request
2. Assign a channel_ord value in request

Using MAM file is the reason that we have to return channel_ord
in resposne, since the user may need to know how much is the
current channel_ord

channel_ord is in trint18_t in entangled. According to MAM spec
trint18_t is a type in the range of +- (3^18-1)/2, so the type of
channel_ord is modified from uint8_t from int32_t.

+++ request/ta_send_mam.h
@@ -27,7 +27,7 @@ extern "C" {
 typedef struct send_mam_req_s {
   tryte_t prng[NUM_TRYTES_ADDRESS + 1];
   char* payload;
-  uint8_t channel_ord;
+  int32_t channel_ord;
 } ta_send_mam_req_t;
 
 /**
